requirements.txt
google-generativeai>=0.6.0
yt-dlp>=2024.12.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.28.0
tqdm>=4.65.0
